kinda like the last commit to this project
did nothing, just gonna drop this project here. I realized while implementing the node class that everything I need for the blockchain is already there, and all else will just be a method to make this code connected to other instances of this code. Maybe I'll get back here when(/if) I learn enough web development to implement the node + GUI framework for this :)